feat: reorganizar importaciones y actualizar componente de visualización de la descripción
